# backend/swagger/swagger.yaml
basePath: /
definitions:
  db.Video:
    properties:
      created_at:
        $ref: '#/definitions/pgtype.Timestamp'
      duration_sec:
        $ref: '#/definitions/pgtype.Int4'
      id:
        type: string
      status:
        $ref: '#/definitions/db.VideoStatus'
      title:
        type: string
      user_id:
        type: string
    type: object
  db.VideoStatus:
    enum:
    - PREUPLOAD
    - UPLOADED
    - PROCESSING
    - READY
    - FAILED
    type: string
    x-enum-varnames:
    - VideoStatusPREUPLOAD
    - VideoStatusUPLOADED
    - VideoStatusPROCESSING
    - VideoStatusREADY
    - VideoStatusFAILED
  pgtype.InfinityModifier:
    enum:
    - 1
    - 0
    - -1
    format: int32
    type: integer
    x-enum-varnames:
    - Infinity
    - Finite
    - NegativeInfinity
  pgtype.Int4:
    properties:
      int32:
        format: int32
        type: integer
      valid:
        type: boolean
    type: object
  pgtype.Timestamp:
    properties:
      infinityModifier:
        $ref: '#/definitions/pgtype.InfinityModifier'
      time:
        description: Time zone will be ignored when encoding to PostgreSQL.
        type: string
      valid:
        type: boolean
    type: object
  server.Asset:
    properties:
      content_type:
        type: string
      href:
        type: string
      id:
        type: string
      name:
        type: string
      original_name:
        type: string
      size:
        type: integer
    type: object
  server.AssetsRequest:
    properties:
      content_type:
        type: string
      name:
        type: string
      size:
        type: integer
    required:
    - content_type
    - name
    - size
    type: object
  server.AssetsResponse:
    properties:
      data:
        $ref: '#/definitions/server.AssetsResponseData'
      error: {}
      message:
        type: string
    type: object
  server.AssetsResponseData:
    properties:
      asset:
        $ref: '#/definitions/server.Asset'
      form:
        additionalProperties:
          type: string
        type: object
      header:
        additionalProperties:
          type: string
        type: object
      upload_url:
        type: string
    type: object
  server.AuthResponse:
    properties:
      error: {}
      message:
        type: string
    type: object
  server.ConfirmSignupRequest:
    properties:
      email:
        type: string
      otp:
        type: string
    required:
    - email
    - otp
    type: object
  server.DatabaseHealthStatus:
    properties:
      acquired_conns:
        type: integer
      idle_conns:
        type: integer
      max_conns:
        type: integer
      status:
        $ref: '#/definitions/server.Status'
      total_conns:
        type: integer
    type: object
  server.GetVideoResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/db.Video'
        type: array
      error: {}
      message:
        type: string
    type: object
  server.HealthResponse:
    properties:
      db:
        $ref: '#/definitions/server.DatabaseHealthStatus'
      status:
        $ref: '#/definitions/server.Status'
    type: object
  server.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  server.Profile:
    properties:
      email:
        type: string
      name:
        type: string
      sub:
        type: string
    type: object
  server.ProfileResponse:
    properties:
      data:
        $ref: '#/definitions/server.Profile'
      error: {}
      message:
        type: string
    type: object
  server.SignUpRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
    required:
    - email
    - name
    - password
    type: object
  server.Status:
    enum:
    - UP
    - DOWN
    type: string
    x-enum-varnames:
    - StatusUp
    - StatusDown
  server.ThumbnailAssetsRequest:
    properties:
      content_type:
        type: string
      name:
        type: string
      size:
        type: integer
    required:
    - content_type
    - name
    - size
    type: object
  server.ThumbnailAssetsResponse:
    properties:
      data:
        $ref: '#/definitions/server.ThumbnailAssetsResponseData'
      error: {}
      message:
        type: string
    type: object
  server.ThumbnailAssetsResponseData:
    properties:
      upload_url:
        type: string
    type: object
  server.UpdateMetadataRequest:
    properties:
      duration_sec:
        type: integer
      status:
        allOf:
        - $ref: '#/definitions/db.VideoStatus'
        enum:
        - PREUPLOAD
        - UPLOADED
        - PROCESSING
        - READY
        - FAILED
      title:
        type: string
      user_id:
        type: string
    type: object
externalDocs:
  description: OpenAPI
  url: https://swagger.io/resources/open-api/
host: localhost:8080
info:
  contact:
    email: subrotokumar@outlook.in
    name: Subroto Kumar
    url: http://github.com/subrotokumar/playstack
  description: API Dccumentation for Glitch Backend.
  license:
    name: Apache 2.0
  title: Playstack
  version: "1.0"
paths:
  /auth/confirm-signup:
    post:
      consumes:
      - application/json
      description: Confirm a user's signup using OTP
      parameters:
      - description: Confirm signup payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/server.ConfirmSignupRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/server.AuthResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server.AuthResponse'
      summary: Confirm signup
      tags:
      - auth
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate a user and set access/refresh cookies
      parameters:
      - description: Login payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/server.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/server.AuthResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/server.AuthResponse'
      summary: Login user
      tags:
      - auth
  /auth/profile:
    post:
      consumes:
      - application/json
      description: Get Profile Detail
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.ProfileResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/server.ProfileResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server.ProfileResponse'
      summary: Profile
      tags:
      - auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Refresh access token using refresh token cookie
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/server.AuthResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server.AuthResponse'
      summary: Refresh access token
      tags:
      - auth
  /auth/signup:
    post:
      consumes:
      - application/json
      description: Create a new user account
      parameters:
      - description: Sign up payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/server.SignUpRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/server.AuthResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server.AuthResponse'
      summary: Sign up a new user
      tags:
      - auth
  /health/liveness:
    get:
      description: Indicates whether the application process is alive
      produces:
      - application/json
      responses:
        "200":
          description: Service is alive
          schema:
            $ref: '#/definitions/server.HealthResponse'
      summary: Liveness probe
      tags:
      - health
  /health/readiness:
    get:
      description: Indicates whether the application is ready to receive traffic
      produces:
      - application/json
      responses:
        "200":
          description: Service is ready or not
          schema:
            $ref: '#/definitions/server.HealthResponse'
      summary: Readiness probe
      tags:
      - health
  /internal/media/videos/{videoId}:
    patch:
      consumes:
      - application/json
      description: Updates title, status, or duration of a video
      parameters:
      - description: Video ID
        in: path
        name: videoId
        required: true
        type: string
      - description: Metadata update payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/server.UpdateMetadataRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/server.AssetsResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server.AssetsResponse'
      security:
      - BasicAuth: []
      summary: Update video metadata (internal)
      tags:
      - media-internal
  /media/videos:
    get:
      description: Returns videos with READY status
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.GetVideoResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server.GetVideoResponse'
      security:
      - BearerAuth: []
      summary: List ready videos
      tags:
      - media
  /media/videos/{videoId}/thumbnail/signed-url:
    put:
      consumes:
      - application/json
      description: Returns a presigned PUT URL for uploading a video thumbnail
      parameters:
      - description: Video ID
        in: path
        name: videoId
        required: true
        type: string
      - description: Thumbnail metadata
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/server.ThumbnailAssetsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.ThumbnailAssetsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/server.AssetsResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/server.AssetsResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server.AssetsResponse'
      security:
      - BearerAuth: []
      summary: Create presigned URL for thumbnail upload
      tags:
      - media
  /media/videos/signed-url:
    post:
      consumes:
      - application/json
      description: Creates a video record and returns a presigned POST URL for uploading
        raw media
      parameters:
      - description: Video asset metadata
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/server.AssetsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.AssetsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/server.AssetsResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server.AssetsResponse'
      security:
      - BearerAuth: []
      summary: Create presigned URL for video upload
      tags:
      - media
securityDefinitions:
  BasicAuth:
    type: basic
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"